Added signing students to signing package, and cleaned up after it in other packages
