+++ src/main/java/com/example/demo/course/CourseService.java
@@ -1,31 +1,24 @@
 package com.example.demo.course;
 
-import com.example.demo.course.exceptions.CourseAlreadyTakenException;
 import com.example.demo.course.exceptions.CourseDoesNotExistException;
 import com.example.demo.course.exceptions.NameTakenException;
 
-import com.example.demo.course.exceptions.TooManyStudentsException;
-import com.example.demo.student.Student;
 import com.example.demo.student.StudentRepository;
-import com.example.demo.student.exceptions.StudentDoesNotExistException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 @Service
 public class CourseService {
 
     private final CourseRepository courseRepository;
-    private final StudentRepository studentRepository;
 
     @Autowired
-    public CourseService(CourseRepository courseRepository, StudentRepository studentRepository) {
+    public CourseService(CourseRepository courseRepository) {
         this.courseRepository = courseRepository;
-        this.studentRepository = studentRepository;
     }
 
     public List<Course> getCourses(){
@@ -66,20 +59,7 @@ public class CourseService {
 
     }
 
-    @Transactional
-    public void signStudent(Long studentId , String courseName){
 
-        Student student = studentRepository.findById(studentId).orElseThrow(() -> new StudentDoesNotExistException(StudentDoesNotExistException.ERROR_THERE_IS_NO_STUDENT_WITH_ID + studentId));
-        Course course = courseRepository.findCourseByName(courseName).orElseThrow(() -> new CourseDoesNotExistException(CourseDoesNotExistException.COURSE_DO_NOT_EXIST));
-        if(course.getStudent().size() >= course.getMaxNumberOfStudents()){
-            throw new TooManyStudentsException(TooManyStudentsException.TOO_MANY_STUDENTS);
-        }else if(!student.getCourses().contains(course)){
-            student.getCourses().add(course);
-        }else{
-            throw new CourseAlreadyTakenException(CourseAlreadyTakenException.COURSE_ALREADY_TAKEN);
-        }
-
-    }
 
     public void checkIfNameTaken(Course course) {
         Optional<Course> courseOptional = courseRepository.findCourseByName(course.getName());
